model/requirements.txt
torch>=2.2
torchvision>=0.17
facenet-pytorch
onnx>=1.15
onnxruntime>=1.17
coremltools>=7.2
tensorflow>=2.14
Pillow
tqdm
numpy
